test: preserve metrics contract shard placement
